requirements.txt
# ArXiv Research Assistant - Requirements
# Core Dependencies
torch>=1.13.1,<2.0.0
transformers>=4.35.0
Pillow>=9.0.1
requests>=2.31.0
numpy>=1.21.0
pandas>=1.3.0
tqdm>=4.62.0

# Search and Storage
faiss-cpu>=1.7.4
# faiss-gpu>=1.7.4  # Uncomment for GPU support

# Optional: RocksDB for metadata storage
# rocksdict>=0.3.0  # Uncomment if using RocksDB

# File Processing
PyPDF2>=3.0.0
arxiv>=1.4.0

# Web Framework (Django)
Django>=4.2.0,<5.0.0
django-cors-headers>=4.0.0

# API Integration
groq>=0.4.0
# openai>=1.0.0  # Uncomment if using OpenAI

# Development and Testing
pytest>=7.0.0
pytest-django>=4.5.0
black>=23.0.0
flake8>=6.0.0

# AWS Lambda (if deploying to Lambda)
# boto3>=1.26.0
# awslambdaric>=2.0.0

# Production
gunicorn>=21.0.0
python-dotenv>=1.0.0

# Optional: For advanced features
# langchain>=0.1.0
# chromadb>=0.4.0
